% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{atan3}
\alias{atan3}
\title{Calculate angle between a vector and y-axis}
\usage{
atan3(x, y)
}
\arguments{
\item{x}{The x coordinate from the (x, y) locations.}

\item{y}{The y coordinate from the (x, y) location for an organism.}
}
\description{
Calculate the angle between a vector and y-axis, not the x-axis as in atan2
}
\details{
This is a helper function. It is called in another function but is otherwise
not "outward facing".

This function calculates the angle between a vector and the y-axis. The
base function atan2 calculates the angle between a vector and the x-axis,
which is not desirable in this case.

If the data will be aggregated, it is recommended to aggregate the data before
running this function.
}
