% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{calcBearing}
\alias{calcBearing}
\title{Calculate bearing}
\usage{
calcBearing(list)
}
\arguments{
\item{list}{A list of data frames with separate columns for x and y
coordinate values.}
}
\value{
A list of data frames with a column for the bearing of the organism
  at each time step.
}
\description{
Calculates the bearing (direction of movement) for each time step
}
\details{
This function requires that the data has been previously processed with the
\code{calcXY()} function, providing (x, y) coordinates. Calculate the
direction moved by the organism relative to the y axis between each time step
in your data frames.

If the data will be aggregated, it is recommended to aggregate the data before
running this function.
}
\examples{
# Provide a list of data frames with two columns for the (x, y) coordinates
\dontrun{dat <- calcBearing(dat)}
}
