% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{calcTurnVelocity}
\alias{calcTurnVelocity}
\title{Calculate turn velocity}
\usage{
calcTurnVelocity(list)
}
\arguments{
\item{list}{A list of data frames, where each data frame has a column for
turn angle.}
}
\value{
A list of data frames that each contain a column for turn velocity.
}
\description{
Calculate the turning velocity in degrees per second between two moves
}
\details{
For this function to work, the data must have previously been processed with
the \code{calcTurnAngle} function.

This function calculates the turning velocity between two consecutive moves.
The units for turn velocity will be degrees per second.

If the data will be aggregated, it is recommended to aggregate the data before
running this function.
}
\examples{
# Provide data previously processed by the calcTurnAngle function
\dontrun{dat <- calcTurnVelocity(dat)}
}
