% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{calcXY}
\alias{calcXY}
\title{Calculate (x, y) coordinates}
\usage{
calcXY(list)
}
\arguments{
\item{list}{A list of data frame objects with columns dx and dy.}
}
\value{
Converts dx and dy values to (x, y) coordinates.
}
\description{
Calculates (x,y) coordinates from dx and dy values
}
\details{
Use the dx and dy columns in the servosphere data frames to calculate the (x,
y) coordinate for each position recording. If the data will be thinned, it is
recommended to thin the data before running this function.
}
\examples{
\dontrun{dat <- calcXY(dat) # Calculate (x, y) coordinate pairs}
}
