% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{servosphere}
\alias{servosphere}
\title{Example servosphere data}
\format{A list of two data frames:
\describe{
    \item{X.cState}{stimulus designator}
    \item{dT..ms.}{Change in time, in miliseconds}
    \item{dx..cm.}{Distance moved in X direction during dT, in centimeters}
    \item{dY..cm.}{Distance moved in Y direction during dT, in centimeters}
}}
\usage{
servosphere
}
\description{
A list containing two data frames of example data output from the software
used with a servosphere.This data has not been cleaned in anyway.
}
\keyword{datasets}
