% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{servosphere_aggregate_ex}
\alias{servosphere_aggregate_ex}
\title{Example trial record}
\format{A data frame:
\describe{
    \item{id}{Unique identifier for each trial}
    \item{treatment}{Treatment factor for each trial}
    \item{date}{Date trial was performed}
    \item{time}{Time trial was performed}
}}
\usage{
servosphere_aggregate_ex
}
\description{
A data frame containing information relevant to experimental treatments for
the example servosphere data, "servosphere".
}
\keyword{datasets}
