% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summaryAvgVelocity}
\alias{summaryAvgVelocity}
\title{Calculate average velocity}
\usage{
summaryAvgVelocity(list, summary.df = NA)
}
\arguments{
\item{list}{A list of data frames, each of which has a column for bearing.}

\item{summary.df}{The data frame object within which you are storing path
  summary variables. The default is NA if you do not currently have a summary
  data frame object started. When set to \code{NA} the function will create a new
summary data frame. When an object is provided, the function will merge the
summary data frame with the new data.}
}
\value{
The inputed summary data frame or a new data frame if summary.df is
  \code{NA}
}
\description{
Calculate the average velocity for a movement path.
}
\details{
Calculate the average velocity for a movement path. The units on velocity are
equal to the distance units used to record the data per second.
}
\examples{
# If a summary data frame has not been started
\dontrun{summary_df <- summaryAvgVelocity(dat, summary.df = NA)}
# If a summary data frame has been started
\dontrun{summary_df <- summaryAvgVelocity(dat, summary.df = summary_df)}
}
