% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trial_id_split}
\alias{trial_id_split}
\title{Example trial record}
\format{A data frame:
\describe{
    \item{id}{Unique identifier for each trial}
    \item{stimulus}{Stimulus identifier}
    \item{treatment}{Treatment factor for each trial}
    \item{date}{Date trial was performed}
    \item{time}{Time trial was performed}
    \item{id_stim}{Unique identifier for each stimulus trial combination}
}}
\usage{
trial_id_split
}
\description{
A data frame containing information relevant to experimental treatments for
the example servosphere data, "servosphere".
}
\keyword{datasets}
