% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived_variables.R
\name{xyCoordinates}
\alias{xyCoordinates}
\title{Calculate (x, y) coordinates}
\usage{
xyCoordinates(list)
}
\description{
Calculates (x,y) coordinates from dx and dy values
}
\details{
Use the dx and dy columns in the servosphere data frames to calculate the (x,
y) coordinate for each position recording. This function can be used before
or after thinning, but it will run faster on thinned data.
}
