% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/servosphereR.R
\docType{package}
\name{servosphereR}
\alias{servosphereR}
\alias{servosphereR-package}
\title{servosphereR: A package to facilitate cleaning and analyzing data produced by a servosphere.}
\description{
A servosphere, or locomotory compensator, is a device that can be used to
study the movement of insects or other small critters. A servosphere consists
of a ball that rests on top of motors. A camera pointed at the top of the
sphere watches the insect as it moves and reports position changes to the
motors. As the insect moves, the motors rotate to keep the insect positioned
at the top of the sphere.
}
