% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variables.R
\name{summaryAvgBearing}
\alias{summaryAvgBearing}
\title{Calculate average bearing}
\usage{
summaryAvgBearing(list, summary.df = NA)
}
\arguments{
\item{list}{A list of data frames, each of which has a column for bearing.}

\item{summary.df}{The summary data frame containing a column recording
the bearing of each recorded movement.}
}
\value{
A list of two named vectors. The first named vector contains the
  average bearing calculated for each movement path. The second named vector
  contains the rho, a measure of concentration for the average bearing.
}
\description{
Calculate the average bearing, or direction of movement, for a movement path.
}
\details{
The bearing is the direction of movement and and falls between 0 and 360.
Outdoors degree measures of 0/360 will typically correspond to due north. For
servosphere data, 0/360 will correspond to the  movement in the direction of
the positive y-axis.
}
